Refactor CodeGen unit tests
