+++ dawn/test/unit-test/dawn/CodeGen/CMakeLists.txt
@@ -12,9 +12,5 @@
 ##
 ##===------------------------------------------------------------------------------------------===##
 
-add_subdirectory(AST)
-add_subdirectory(Optimizer)
-add_subdirectory(IIR)
-add_subdirectory(SIR)
-add_subdirectory(Support)
-add_subdirectory(CodeGen)
+add_subdirectory(Cuda)
+add_subdirectory(Naive)
